Add the game over state and main game loop
Once the timer reaches zero, the game ends and switches the state to the
GameOver state. Phaser allows an argument to be passed between state
switches, and the round's score will be passed to show the player at the
end of their game.
A placeholder exists for highscores, which will involve the localstorage
API at a later time. Currently score isn't functioning properly, as the
tower's physics needs to be tweaked to "feel" right.
Added 'init' method to mainGame so that resetting the time works;
without it the time stays at zero and immediately ends the game when the
reset button is pressed.